# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/orders/status": {
            "put": {
                "description": "Update the status of an order to track its progress through the fulfillment pipeline. Status changes help coordinate between vendors, delivery partners, and customers.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Update Order Status",
                "parameters": [
                    {
                        "description": "Order status update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/update_order_status.UpdateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order status updated successfully",
                        "schema": {
                            "$ref": "#/definitions/update_order_status.UpdateOrderStatusResponse"
                        }
                    }
                }
            }
        },
...
        "/api/vendor/order-summary": {
            "get": {
                "description": "Retrieve order count statistics for the current day, grouped by order status. Provides a quick overview of vendor's daily order performance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order-analytics"
                ],
                "summary": "Get Vendor Today's Order Summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Today's order summary retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/vendors.TodayOrderSummary"
                        }
                    }
                }
            }
        },
        "/api/vendor/orders": {
            "get": {
                "description": "Retrieve a paginated list of orders for a specific vendor. Maximum 10 orders per page to ensure optimal performance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get Vendor Orders List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10, max: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/vendors.OrdersListResponse"
                        }
                    }
                }
            }
        },
...
        "/api/vendors/{vendor_id}/inventory/{item_id}": {
            "put": {
                "description": "Update stock quantity, availability status, or price override for an existing inventory item. All fields are optional - only provide fields you want to update.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Update Inventory Item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID to update",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Inventory update request (all fields optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vendors.UpdateInventoryRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Permanently remove an item from the vendor's inventory. This action cannot be undone. The item will no longer be available for sale.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Remove Item from Vendor Inventory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID to remove from inventory",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "update_order_status.UpdateOrderStatusRequest": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "update_order_status.UpdateOrderStatusResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "vendors.OrderListItem": {
            "type": "object",
            "properties": {
                "customer_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "delivered_by_time": {
                    "type": "string",
                    "example": "2025-05-30T15:00:00Z"
                },
                "order_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "order_status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "accepted",
                        "packed",
                        "ready",
                        "completed",
                        "cancelled",
                        "rejected"
                    ],
                    "example": "packed"
                },
                "order_time_placed": {
                    "type": "string",
                    "example": "2025-05-30T13:30:00Z"
                },
                "pack_by_time": {
                    "type": "string",
                    "example": "2025-05-30T14:30:00Z"
                },
                "pick_by_time": {
                    "type": "string",
                    "example": "2025-05-30T14:45:00Z"
                },
                "total_amount": {
                    "type": "number",
                    "example": 45.99
                }
            }
        },
        "vendors.OrdersListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vendors.OrderListItem"
                    }
                },
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_prev": {
                    "type": "boolean",
                    "example": false
                },
                "limit": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 1,
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total_count": {
                    "type": "integer",
                    "example": 125
                },
                "total_pages": {
                    "type": "integer",
                    "example": 13
                }
            }
        },
        "vendors.TodayOrderSummary": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "integer",
                    "example": 8
                },
                "cancelled": {
                    "type": "integer",
                    "example": 3
                },
                "completed": {
                    "type": "integer",
                    "example": 45
                },
                "packed": {
                    "type": "integer",
                    "example": 15
                },
                "pending": {
                    "type": "integer",
                    "example": 12
                },
                "ready": {
                    "type": "integer",
                    "example": 6
                },
                "rejected": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "vendors.UpdateInventoryRequest": {
            "type": "object",
            "properties": {
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "price_override": {
                    "type": "number",
                    "minimum": 0,
                    "example": 10.99
                },
                "stock_quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 30
                }
            }
        },
...
    }
}