# src/vendor_ec2/update_order_status/req_model.go
package update_order_status

// Valid order statuses
var ValidOrderStatuses = map[string]bool{
	"pending":   true,
	"accepted":  true,
	"packed":    true,
	"ready":     true,
	"completed": true,
	"cancelled": true,
	"rejected":  true,
}

// Request/Response structures
type UpdateOrderStatusRequest struct {
	OrderID string `json:"order_id"`
	Status  string `json:"status"`
}

type UpdateOrderStatusResponse struct {
	Success bool   `json:"success"`
	Message string `json:"message"`
}